vm: implementation that supports account abstraction model (includes signle key wallet) (#3229)
related https://github.com/spacemeshos/go-spacemesh/issues/3220

- state

account extend with Template and State fields.
stub accounts will have both Template and State as nil. When account is spawned with template - Template will store the address of the template, and State will store mutable and immutable state encoded using scale.

- address

when account is spawned address is computed as `sha256(template_address, spawn_args)`. this must be taken into an account when configuring the coinbase for a node.

- validation and mempool

tx id is computed as a `sha256(raw_tx)`, note that raw_tx is a full body of the transaction including signature.
tx header can't be decoded without executing template code (specifically parse_payload and max_spend methods). 

in the upcoming pr - mempool will call into vm to initialize a validation flow (which includes parsing of the tx header). validation api is a 2 step process:
- parse. returns transaction header that can be validated against conservative state
- verify. executes templates verify method. it is assumed that this method costs much more than parse, so it makes sense to call it only if conservative validation succeeded.

- applying transaction

vm.Apply API method accepts layer id, transactions and rewards. Every transaction loads state, executes parse payload, initializes state and executed. Template/Handler API is specified in `genvm/core/types.go`.

Transaction can be skipped for 3 reasons:
- transaction is malformed (expectation is that such transactions should not be added to the block at all)
- account is not spawned (same as a nonce mismatch)
- nonce doesn't match the expected nonce of the account
- can't cover max gas with the account balance

If transaction is skipped it is returned back to the caller (presumably conservative state and the caller is free to do anything about it). Otherwise transaction is consumed and will pay fee to smeshers (in the followup pr for every consumed transaction we will store execution result). Logic for the fee distribution is not changed.