tortoise: ensure that blocks and ballots are processed after ref height was computed (#3494)
reference height is computed when the first layer of the epoch is terminated
before that all ballots and blocks that are passed to OnBlock/OnBallot should be ignored, 
processing them will lead to inconsistent state. namely ballots wont have votes for all blocks in the previous layers 

if there a gap in the layer that were sent to tortoise (due to the failed hare), tortoise waits only for zdist until moving on.
once it decided to move on it will download all relevant data (blocks and ballots) after ref height was computed.