vm: add nonce to self-spawn (#3469)
closes: https://github.com/spacemeshos/go-spacemesh/issues/3392

self-spawn transaction can fail if it was included into the block and account can't cover
gas required for it to be executed (in general it is not possible if all assumptions are satisfied)

however if it does happen there should be a way to retry self-spawning account, so it will have a nonce
as any other transaction. 

in system tests we don't expect self-spawn to fail, so in that case self-spawn submitted always with zero nonce.
if it fails it is definitely a bug and needs to be fixed.