tortoise: not using base ballots with abstain votes for empty layers (#3513)
closes: https://github.com/spacemeshos/go-spacemesh/issues/3446

to encode votes tortoise picks base ballot and encodes voting diff (support, against for blocks, abstain for layers) explicitly. if hare didn't terminate with empty layer immediately then tortoise will vote abstain for that layer within zdist. unless it receives certificate and updates local opinion it should vote against that layer after zdist layers passed.

it is obvious that we can't encode against votes for an empty layer explicitly, because it has no blocks and there is nothing to add to the diff. therefore we should not pick base ballot that votes abstain for a layer if it is empty and already terminated by hare.

this change also updates hare to not output empty block in case of failures, as discussed in the issue above.