Avoid scale code generation for ActivationTxHeader (#3549)
## Motivation
I noticed that a file was changed after running `make generate`. It seems that the code was not regenerated after changes introduced in #3519. It actually turns out that scale (de)serialization code should not be generated for it (https://github.com/spacemeshos/go-spacemesh/pull/3549#issuecomment-1243936342).

## Changes
Moved `ActivationTxHeader` to a separate file.

## Test Plan
unittests, systests

## DevOps Notes
<!-- Please uncheck these items as applicable to make DevOps aware of changes that may affect releases -->
- [x] This PR does not require configuration changes (e.g., environment variables, GitHub secrets, VM resources)
- [x] This PR does not affect public APIs
- [x] This PR does not rely on a new version of external services (PoET, elasticsearch, etc.)
- [x] This PR does not make changes to log messages (which monitoring infrastructure may rely on)