Improve checks in grpcserver_test for better failure messages (#3613)
## Motivation
Currently, when the tests in the `grpcserver_test.go` file fail (which happen sporadically - the tests are flaky), they emit an enigmatic log:

> Error:          inner account data item has wrong data type

That is not super helpful. A test could print what type it actually got vs the exptected one.

## Changes
Use `require.IsType(...)` to check if `foo interface{}` is of the expected type. It will print the actual type of the `foo` in case of failure.

## DevOps Notes
<!-- Please uncheck these items as applicable to make DevOps aware of changes that may affect releases -->
- [x] This PR does not require configuration changes (e.g., environment variables, GitHub secrets, VM resources)
- [x] This PR does not affect public APIs
- [x] This PR does not rely on a new version of external services (PoET, elasticsearch, etc.)
- [x] This PR does not make changes to log messages (which monitoring infrastructure may rely on)