Generalize storing of key value pairs in DB (#3616)
## Motivation
Pull out some of the changes from https://github.com/spacemeshos/go-spacemesh/pull/3567 for easier review.

## Changes
The activation package needs to store transient data in the DB that is needed to create ATXs for the time between syncing global state and publishing a new ATX, after which this data can be discarded.

At the moment this is done for `NIPostBuilderState` and `NIPostChallange`. The changes in this PR generalize this storing, retrieving and clearing of data. In https://github.com/spacemeshos/go-spacemesh/pull/3567 another datatype is added to the `kvstore` introduced in this PR.

## Test Plan
Existing tests pass, new code is sufficiently tested.

## TODO
<!-- This section should be removed when all items are complete -->
- [X] Explain motivation or link existing issue(s)
- [x] Test changes and document test plan
- [x] Update documentation as needed

## DevOps Notes
<!-- Please uncheck these items as applicable to make DevOps aware of changes that may affect releases -->
- [x] This PR does not require configuration changes (e.g., environment variables, GitHub secrets, VM resources)
- [x] This PR does not affect public APIs
- [x] This PR does not rely on a new version of external services (PoET, elasticsearch, etc.)
- [x] This PR does not make changes to log messages (which monitoring infrastructure may rely on)