tortoise: tally votes for layers independently of the hare (#3596)
depends on: https://github.com/spacemeshos/go-spacemesh/issues/3569
closes: https://github.com/spacemeshos/go-spacemesh/issues/3575
closes: https://github.com/spacemeshos/go-spacemesh/issues/2919

1. onBlock and onBallot refactored in a way to avoid failures if required state wasn't loaded and to support counting ballots separately
 
onBlock will not fail if reference height wasn't computed yet. reference height is always computed together with the first layer in the epoch. this change makes it safe to call onBlock before such first layer was processed.

onBallot will not fail if beacon is not yet available. beacon needs to be known only before counting ballot (in both modes). this change (and changes to onBlock) allows to decode votes before calling a method to verify a layer,  in practice decoding votes will be the most expensive operation. with an exception of counting votes in full mode.

this refactoring also allows to seamlessly count votes from late ballot without any additional logic in place, this will be required for https://github.com/spacemeshos/go-spacemesh/issues/3573.

2. new TallyVotes method

this method is safe to call whenever node needs to know the latest local opinion, in current codebase it is called:
- to check that local opinion is consistent with previously applied blocks
- before casting votes. this ensures that the node always votes according to the latest changes in the opinion