Fix race condition in Builder::StartSmeshing (#3659)
## Motivation
There is a race condition in `StartSmeshing()` happening when it is called repeatedly with `StopSmeshing()` (not even concurrently). The race is on the atomic boolean `started`. It is set to false in the spawned goroutine (which is fine) and also in `StopSmeshing()` (which is not fine because it no longer serves as a signalling mechanism that Builder stopped smeshing).

Consider the following:

1. `StartSmeshing()`
   1. `started` is set to true
   2. goroutine is spawned
2. `StopSmeshing()`
   1. `started` is set to false. At this point, one can already restart smeshing
   2. b.stop() - cancel context, the goroutine starts to shutdown
3. `StartSmeshing()`
   1. `started` is set to true
   2. goroutine finishes. It sets `started` back to false. Builder state is broken now. It is smeshing, but `Smeshing()` returns false.

Closes #3066 

## Changes
Removed setting `started = false` in `StopSmeshing()`. Added more tests and fixed existing ones

## Test Plan
Unit tests. Tests now pass with `-race` flag.

## DevOps Notes
- [x] This PR does not require configuration changes (e.g., environment variables, GitHub secrets, VM resources)
- [x] This PR does not affect public APIs
- [x] This PR does not rely on a new version of external services (PoET, elasticsearch, etc.)
- [x] This PR does not make changes to log messages (which monitoring infrastructure may rely on)