Run systests in Job (#3730)
## Motivation
System tests in CI fail sporadically because `kubectl logs -f <systest pod>` exits prematurely (while tests are still running). An example of a CI run failed in this way: https://github.com/spacemeshos/go-spacemesh/actions/runs/3421691184/jobs/5698440170.

I suspect this is because of log rotation, which `kubectl` doesn't support properly (see https://github.com/kubernetes/kubernetes/issues/59902 and https://github.com/kubernetes/kubernetes/issues/28369).

## Changes
The fix is to wait for the systests to actually finish. It turns out that it's not possible when using a Pod because `kubectl wait --for=<>` doesn't provide a way to wait for Pod to "complete". Kubernetes has a different kind of object for this - a [Job](https://kubernetes.io/docs/concepts/workloads/controllers/job/).

Hence in this PR, I changed the deployment of systests to run in Jobs and I added a script to wait for the tests to finish.

`make run` now returns the exit code properly depending if the tests passed or failed. Thanks to it the CI step running tests is automatically marked as failed if tests failed - I could remove additional steps to check tests result.

Finally, in order to spawn a Job with restart policy = Never and additional labels, I had to create it via `kubectl apply -f` because `kubectl create job` is too limited. In order to create a Job deployment yaml file (to be used with `kubectl apply -f`), I use a template file (`systest_job.yml.tmpl`) which is rendered using [gomplate](https://gomplate.ca).

## Test Plan
CI passing

## DevOps Notes
- [x] This PR does not require configuration changes (e.g., environment variables, GitHub secrets, VM resources)
- [x] This PR does not affect public APIs
- [x] This PR does not rely on a new version of external services (PoET, elasticsearch, etc.)
- [x] This PR does not make changes to log messages (which monitoring infrastructure may rely on)