Modify wait time in waitForFirstATX (#3721)
## Motivation
Closes #3698 - see it for details.

## Changes
### Poet proof deadline adjusted
Previously the deadline to wait for poet proofs was set as late as possible. As a result, the nodes created their ATXes very late (close to the start of the next poet round). It caused problems for the nodes that are new to the network.

This PR shift the deadline to the earliest possible moment which is "poet round end" + grace period.

## Time to wait for the first ATX extended
Nodes that just joined the network need a valid, "fresh", Positioning ATX targeting the current epoch. At the start, they wait a specific time to obtain it. This PR extends this wait time to wait for Poet round END + 2 * grace period + "average post duration" (hardcoded).

:bulb: I'm not quite sure about the hardcoded average post duration. It's meant to accommodate the time that other nodes need to spend generating a PoST proof to build their ATXes. I'm not sure what value it should have nor if it is really needed. 2*Grace period should be enough anyway.

## Test Plan
systests (especially `TestAddNodes`)

## DevOps Notes
- [x] This PR does not require configuration changes (e.g., environment variables, GitHub secrets, VM resources)
- [x] This PR does not affect public APIs
- [x] This PR does not rely on a new version of external services (PoET, elasticsearch, etc.)
- [x] This PR does not make changes to log messages (which monitoring infrastructure may rely on)