migrate-ci-cluster (#3878)
## Motivation
<!-- Please mention the issue fixed by this PR or detailed motivation -->
We need to migrate Github Actions Tests to this new GKE Infrastructure because current cluster is having random kernel panics on the nodes
Closes #
<!-- `Closes #XXXX, closes #XXXX, ...` links mentioned issues to this PR and automatically closes them when this it's merged -->

## Changes
<!-- Please describe in detail the changes made -->
I add 4 new environment variables to consume from Github Repo secrets. This variables allow the repo to connect to the new GKE infrastructure to deploy in the nodepool GHA the jobs.

## Test Plan
<!-- Please specify how these changes were tested 
(e.g. unit tests, manual testing, etc.) -->
We tests the new jobs running in the new cluster. If this fails, we can restore back the line
```yaml
        run: gcloud container clusters get-credentials ${{ secrets.CI_CLUSTER_NAME }} --region ${{ secrets.CI_REGION_NAME }} --project ${{ secrets.CI_GCP_PROJECT_ID }}

```

## DevOps Notes
<!-- Please uncheck these items as applicable to make DevOps aware of changes that may affect releases -->
- [:fire:] This PR does not require configuration changes (e.g., environment variables, GitHub secrets, VM resources)
- [x] This PR does not affect public APIs
- [x] This PR does not rely on a new version of external services (PoET, elasticsearch, etc.)
- [x] This PR does not make changes to log messages (which monitoring infrastructure may rely on)