Use multierror pkg for returning mutliple errors as `error` (#3881)
## Motivation
The [multierror](https://github.com/hashicorp/go-multierror) pkg greatly simplifies returning a collection of errors (`[]error`). It provides an `Error` type that supports appending many errors and implements `Error` interfaces and `Unwrap()`.

It also provides a `Group` that works similarly to `errgroup.Group`, but aggregates returned errors. It's useful when we want all goroutines to continue in spite of one of them failing **and** collect all the errors.

## Changes
- use `multierror.Error` where `[]error` was returned,
- use `multierror.Group` where errors from goroutines need to be aggregated.