ATX Builder early detection of stale challenges. (#3924)
## Motivation
When the ATX Builder tries to publish an ATX, it first creates a challenge that commits to a certain publication layer. It might happen that the challenge is not technically stale _yet_ (the `PubLayerID` is in the future), but it's impossible to create a poet proof on time as the respective poet rounds have already started. In such a case, it makes no sense to submit that challenge to poets. The builder should detect it and create a fresh challenge targeting the next epoch.

See #3917 where it happens.

## Changes
Builder verifies if the poet round, which would yield proof on time to be able to publish ATX in the requested layer (by `challenge.PubLayerID`), has not started yet.

It allows discarding stale challenges early, before prompting Poets.

Additionally, removed `challenge` from Builder's state. It's now only keeped as a local var and persisted in DB for recovery.

## Test Plan
- unit tests