Proper poet challenge submit timeout (#3999)
## Motivation
Currently, there is a hardcoded 10s timeout for all requests to the Poet service. It's limiting, we might want to wait longer or shorter depending on the situation. The user should set the timeout. 

## Changes
Removed the context factory from the HTTP Poet client. The user of the client decides about the timeout on a per-request basis.

## Test Plan
- added unit tests verifying if a hanged challenge submit times out properly.