Use AwaitLayer instead of subscribing to layer updates for status updates (#4014)
## Motivation
Part of https://github.com/spacemeshos/go-spacemesh/issues/4010. Refactor code to not subscribe to layer changes but rather await new layers.

## Changes
Use `AwaitLayer` in `SubscribeToLayers` instead of a layer timer and make sure that the correct layer is reported / logged. Before if the node was hibernating or busy for some time it might report on multiple layers in a short period. Now it will skip reporting status updates to the current layer to avoid spamming clients with (outdated) updates.

## Test Plan
n/a

## TODO
<!-- This section should be removed when all items are complete -->
- [x] Explain motivation or link existing issue(s)
- [x] Test changes and document test plan
- [x] Update documentation as needed

## DevOps Notes
<!-- Please uncheck these items as applicable to make DevOps aware of changes that may affect releases -->
- [x] This PR does not require configuration changes (e.g., environment variables, GitHub secrets, VM resources)
- [x] This PR does not affect public APIs
- [x] This PR does not rely on a new version of external services (PoET, elasticsearch, etc.)
- [x] This PR does not make changes to log messages (which monitoring infrastructure may rely on)