README: Fix instructions to run go outside of make
Readme instructions updated to be inline with the Makefile. Changed
print-env and print-test-env to print the variable assignments on one
line, this allows them to be easily incorporated into shell commands.

Since the print output could now easily be passed to go env, removed the
go-env and go-test-env targets to reduce duplication.