Use ed25519-recovery instead of old ed25519 library (#4101)
## Motivation
The old ed25519 package is about to be archived. A new `ed25519-recovery` package was created with significant performance improvements and clearer separation of official Go code and our modifications of it as well as increased test coverage. See https://github.com/spacemeshos/ed25519-recovery for more details

This PR migrates go-spacemesh to use the new library over the old one.

## Changes
- replace `spacemeshos/ed25519` with `spacemeshos/ed25519-recovery`

## Test Plan
n/a

## TODO
<!-- This section should be removed when all items are complete -->
- [x] Explain motivation or link existing issue(s)
- [x] Test changes and document test plan
- [x] Update documentation as needed

## DevOps Notes
<!-- Please uncheck these items as applicable to make DevOps aware of changes that may affect releases -->
- [x] This PR does not require configuration changes (e.g., environment variables, GitHub secrets, VM resources)
- [x] This PR does not affect public APIs
- [x] This PR does not rely on a new version of external services (PoET, elasticsearch, etc.)
- [x] This PR does not make changes to log messages (which monitoring infrastructure may rely on)