Fix module level logging
The app log was erroneously being re-assigned using app.addLogger, which
if no specific level was provided for the given module would default to
info. So in this case the app log was being set to an info level log,
even though it was hardcoded to debug at construction time.

The result was that no module level log could set levels lower than
info.

This change simply deletes the re-assignment, and deletes the unused
AppLogger constant.