Reduce wait times
The shared round helps ensure that nodes stay in sync with respect to
hare rounds this means we can be much less lenient with the timing.

However there is still a delay between the broker processing a message
and hare processing it.

Really we should hook into hare to solve this.