Reduce expected leaders
Previously I had adjused this value thinking that it affected the result
of the pre-round, but in fact it is used for the proposal round which
is 2 rounds after the pre-round. It seems likely that the failures I was
seeing were as a consequence of nodes progressing layers faster than
other nodes resulting in trailing nodes dropping messages for future
layers.

I've set the value such that it doesn't look like any BFT threshold,
which can be misleading.