vm: set fixed gas according to pricelist (#4289)
closes: https://github.com/spacemeshos/go-spacemesh/issues/4270

there are three different gas methods - base, load, exec.
base covers verification and intrinsic cost for transaction (different for regular/spawn transactions). base plus cost of the tx data is a total intrinsic cost of the transaction, if account can't cover this cost - we will not run verification, and transaction will be ineffective.

load cost is a cost of loading principal account from storage, it is separated from exec cost so that it can be charged additionally when executing regular spawn.

exec cost is a cost of executing a transaction, includes cost for loading/updating state and fixed price for account access.