Ensure that start smeshing cannot fail silently
This change exports the PrepareInitializer method of the
PostSetupManager so that it can be called before initiating the long
running process of pos initialization. This allows for the start
smeshing api call to directly return an error if PrepareInitializer
fails.

Also this change updates Builder.StartSmeshing to panic if the smeshing
operation fails, this ensures that the node will crash if there is any
problem with smeshing.