Ensure start smeshing cannot fail silently (take 2) (#4338)
## Motivation

Closes https://github.com/spacemeshos/go-spacemesh/issues/4292

## Changes

This change moves the  `prepareInitializer` method of `PostSetupManager.StartSession` and exports it, putting in place the requirement to call `PrepareInitializer` before calling `StartSession`. This provides a means for POST opts to be validated synchronously before starting the long running task of `StartSession` which allows the start smeshing api call to directly return an error to the caller if there is a problem with the smeshing options.

Also this change updates `Builder.StartSmeshing` to panic if the smeshing operation fails, this ensures that the node will crash if there is any problem with the smeshing background task. While not optimal in terms of user experience this at least means that users will now be aware of problems with the smeshing process.

Note: This replaces #4319 which suffered from the problem that a call to `StartSmeshing`, while initialization was active would crash the node. Now `PrepareInitializer` acts as a gate and will instead return an error if the node is initializing, which prevents calls to `StartSession` while the node is initializing.

## Test Plan
Updated existing tests and added a test to check that `PrepreInitializer` returns an error if given invalid config and a test to check that `PrepareInitialzer` and `StartSession` return errors if not called in sequence.

I have manually verified that if `StartSession` returns an error the `activation.Builder` will panic, but I can't write an automated test for it since the go routine that calls `StartSession` is initiated by the `activation.Builder` and there's not a way to neatly insert a recover statement there just for tests.

## DevOps Notes
<!-- Please uncheck these items as applicable to make DevOps aware of changes that may affect releases -->
- [x] This PR does not require configuration changes (e.g., environment variables, GitHub secrets, VM resources)
- [x] This PR does not affect public APIs
- [x] This PR does not rely on a new version of external services (PoET, elasticsearch, etc.)
- [x] This PR does not make changes to log messages (which monitoring infrastructure may rely on)