activation: reduce variance and make poet retries shorter (#4371)
we are using LinearJitterBackoff from retryablehttp. it selects delay in a following way:

```
jitter := rand.Float64() * float64(max-min)
jitterMin := int64(jitter) + int64(min)
return time.Duration(jitterMin * int64(attemptNum))
```

i think the idea to use GracePeriod / 100 as a min and GracePeriod / 10 wasn't the best. as it makes delay greatly variable.
practically the minimum is ~0 (GracePeriod / 100) so the delay is a random fraction of GracePeriod / 10 multiplied by number of the attempts.

for example if grace period is 10s, the worst case delays for consecutive attempts are 1s, 2s, 3s and so on. related to https://github.com/spacemeshos/go-spacemesh/issues/4367 
10m, - 1m, 2m, 3m.. this makes it very likely to wait very long after several failures. this is very likely why we see something like that https://github.com/spacemeshos/go-spacemesh/issues/4363 

closes: https://github.com/spacemeshos/go-spacemesh/issues/4367 
closes: https://github.com/spacemeshos/go-spacemesh/issues/4363